+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
-# mongo-for-intellij Changelog
+# Mongo DB Browser Changelog
 
 ## [Unreleased]
 
@@ -10,4 +10,6 @@
 
 - Initial scaffold created
   from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+- Add Mongo DB Browser and Editor functionality